Update `Topics.jsx`
- Change file extension to `.tsx`
- Move to folder as the component was divided in separate components
